Move LDAP to being an import and refactor the userQuota update
